moving data related functions to data.py and everything else to tools
